backend/pyproject.toml
[project]
name = "canvasflow"
version = "0.1.0"
description = "AI 驱动的图片生成创作平台"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "langchain-core>=0.3.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.3.0",
    "pydantic>=2.7.4",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "aiomysql>=0.2.0",
    "minio>=7.2.0",
    "requests>=2.31.0",
    "httpx>=0.25.0",
    "aiofiles>=23.0.0",
    "pillow>=10.4.0",
    "python-multipart>=0.0.6",
    "alembic>=1.13.0",
    "cryptography>=42.0.0",
    "bcrypt>=4.2.0",
    "pyjwt>=2.9.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/canvasflow"]
